Exercise 10 done

+++ modulo0/ex09/average.c
@@ -2,8 +2,8 @@ int average(int n1, int n2){
 	return (n1+n2)/2;
 }
 
-int average_array(int v [], int n){
-	int sum = 0;
+double average_array(int v [], int n){
+	double sum = 0;
 	for(int i=0; i < n; i++){
 		sum += v[i]; 
 	}
